requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
loguru>=0.5.3
httpx>=0.26.0
backoff>=2.1.0
supabase>=1.0.3
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
google-generativeai>=0.3.2
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.4
langchain-google-genai>=0.0.6
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
python-multipart>=0.0.9
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
flower>=2.0.1
gunicorn>=21.2.0
prometheus-client>=0.19.0
sentry-sdk>=1.40.0
structlog>=24.1.0
tenacity>=8.2.3
typing-extensions>=4.9.0
pyyaml>=6.0.1
jinja2>=3.1.3
aiohttp>=3.9.3
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
python-slugify>=8.0.1
pillow>=10.2.0
python-magic>=0.4.27
boto3>=1.34.0
moto>=4.2.13
pytest-mock>=3.12.0
pytest-env>=1.1.3
pytest-freezegun>=0.4.2
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
bandit>=1.7.7
safety>=2.3.5
pre-commit>=3.6.0 